+++ src/main/java/de/fraunhofer/isst/configmanager/api/controller/UtilController.java
@@ -1,8 +1,8 @@
 package de.fraunhofer.isst.configmanager.api.controller;
 
 import de.fraunhofer.isst.configmanager.api.UtilApi;
-import de.fraunhofer.isst.configmanager.connector.clients.DefaultConnectorClient;
 import de.fraunhofer.isst.configmanager.api.service.UtilService;
+import de.fraunhofer.isst.configmanager.connector.clients.DefaultConnectorClient;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;