+++ src/main/java/de/fraunhofer/isst/configmanager/api/service/ResourceService.java
@@ -1,22 +1,17 @@
 package de.fraunhofer.isst.configmanager.api.service;
 
-import de.fraunhofer.iais.eis.AppRouteImpl;
-import de.fraunhofer.iais.eis.BaseConnector;
-import de.fraunhofer.iais.eis.ContractOffer;
-import de.fraunhofer.iais.eis.DigitalContent;
-import de.fraunhofer.iais.eis.Language;
-import de.fraunhofer.iais.eis.RepresentationImpl;
-import de.fraunhofer.iais.eis.Resource;
-import de.fraunhofer.iais.eis.ResourceBuilder;
-import de.fraunhofer.iais.eis.ResourceImpl;
-import de.fraunhofer.iais.eis.RouteStep;
-import de.fraunhofer.iais.eis.RouteStepImpl;
+import com.fasterxml.jackson.core.JsonProcessingException;
+import com.fasterxml.jackson.databind.JsonNode;
+import com.fasterxml.jackson.databind.ObjectMapper;
+import de.fraunhofer.iais.eis.*;
+import de.fraunhofer.iais.eis.util.RdfResource;
 import de.fraunhofer.iais.eis.util.TypedLiteral;
 import de.fraunhofer.iais.eis.util.Util;
 import de.fraunhofer.isst.configmanager.connector.clients.DefaultConnectorClient;
 import de.fraunhofer.isst.configmanager.connector.clients.DefaultResourceClient;
 import de.fraunhofer.isst.configmanager.model.configlists.EndpointInformationRepository;
 import de.fraunhofer.isst.configmanager.model.endpointinfo.EndpointInformation;
+import de.fraunhofer.isst.configmanager.model.usagecontrol.Pattern;
 import de.fraunhofer.isst.configmanager.util.CalenderUtil;
 import lombok.AccessLevel;
 import lombok.experimental.FieldDefaults;
@@ -30,6 +25,7 @@ import java.net.URI;
 import java.util.ArrayList;
 import java.util.Collection;
 import java.util.List;
+import java.util.Objects;
 
 /**
  * Service class for managing resources.
@@ -454,7 +450,7 @@ public class ResourceService {
 
     /**
      * @param routeSteps  list of route steps
-     * @param visited list of route steps already managed
+     * @param visited     list of route steps already managed
      * @param newResource new resource old version should be replaced with
      */
     private void updateResourceInSubroutes(final List<RouteStep> routeSteps,
@@ -551,7 +547,7 @@ public class ResourceService {
 
     /**
      * @param routeSteps list of route steps
-     * @param visited list of route steps already visited
+     * @param visited    list of route steps already visited
      * @param resourceId id of the resource
      * @return resource
      */
@@ -580,5 +576,193 @@ public class ResourceService {
         }
 
         return foundResource;
+    }
+
+    /**
+     * @param pattern      pattern to create appropriate contract offer
+     * @param contractJson the request body which holds the necessary information
+     * @return contract offer
+     */
+    public ContractOffer getContractOffer(Pattern pattern, String contractJson) throws JsonProcessingException {
+
+        ContractOffer contractOffer = null;
+        switch (pattern) {
+            case PROVIDE_ACCESS:
+                contractOffer = new ContractOfferBuilder()
+                        ._permission_(Util.asList(new PermissionBuilder()
+                                ._title_(Util.asList(new TypedLiteral("Usage Policy")))
+                                ._description_(Util.asList(new TypedLiteral("provide-access")))
+                                ._action_(Util.asList(Action.USE))
+                                .build()))
+                        .build();
+                break;
+            case PROHIBIT_ACCESS:
+                contractOffer = new ContractOfferBuilder()
+                        ._prohibition_(Util.asList(new ProhibitionBuilder()
+                                ._title_(Util.asList(new TypedLiteral("Usage Policy")))
+                                ._description_(Util.asList(new TypedLiteral("prohibit-access")))
+                                ._action_(Util.asList(Action.USE))
+                                .build()))
+                        .build();
+                break;
+            case N_TIMES_USAGE:
+                if (contractJson != null && !contractJson.equals("{}")) {
+                    final var jsonNode = getJsonNodeFromContract(contractJson);
+
+                    String operator = jsonNode.get("binaryoperator").asText();
+                    BinaryOperator binaryOperator;
+                    if (operator.equals("<")) {
+                        binaryOperator = BinaryOperator.LT;
+                    } else if (operator.equals("<=")) {
+                        binaryOperator = BinaryOperator.LTEQ;
+                    } else binaryOperator = BinaryOperator.EQ;
+
+                    String number = jsonNode.get("number").asText();
+                    String pipEndpoint = jsonNode.get("pipendpoint").asText();
+
+                    contractOffer = new NotMoreThanNOfferBuilder()
+                            ._permission_(Util.asList(new PermissionBuilder()
+                                    ._title_(Util.asList(new TypedLiteral("Usage Policy")))
+                                    ._description_(Util.asList(new TypedLiteral("n-times-usage")))
+                                    ._action_(Util.asList(Action.USE))
+                                    ._constraint_(Util.asList(new ConstraintBuilder()
+                                            ._leftOperand_(LeftOperand.COUNT)
+                                            ._operator_(binaryOperator)
+                                            ._rightOperand_(new RdfResource(number, URI.create("xsd:double")))
+                                            ._pipEndpoint_(
+                                                    URI.create(pipEndpoint))
+                                            .build()))
+                                    .build()))
+                            .build();
+                    break;
+                }
+            case DURATION_USAGE:
+                if (contractJson != null && !contractJson.equals("{}")) {
+                    final var jsonNode = getJsonNodeFromContract(contractJson);
+                    String number = jsonNode.get("number").asText();
+
+                    contractOffer = new ContractOfferBuilder()
+                            ._permission_(Util.asList(new PermissionBuilder()
+                                    ._title_(Util.asList(new TypedLiteral("Usage Policy")))
+                                    ._description_(Util.asList(new TypedLiteral("duration-usage")))
+                                    ._action_(Util.asList(Action.USE))
+                                    ._constraint_(Util.asList(new ConstraintBuilder()
+                                            ._leftOperand_(LeftOperand.ELAPSED_TIME)
+                                            ._operator_(BinaryOperator.SHORTER_EQ)
+                                            ._rightOperand_(new RdfResource(number, URI.create("xsd:duration")))
+                                            .build()))
+                                    .build()))
+                            .build();
+                    break;
+                }
+            case USAGE_NOTIFICATION:
+                if (contractJson != null && !contractJson.equals("{}")) {
+                    final var jsonNode = getJsonNodeFromContract(contractJson);
+                    String url = jsonNode.get("url").asText();
+
+                    contractOffer = new ContractOfferBuilder()
+                            ._permission_(Util.asList(new PermissionBuilder()
+                                    ._title_(Util.asList(new TypedLiteral("Usage Policy")))
+                                    ._description_(Util.asList(new TypedLiteral("usage-notification")))
+                                    ._action_(Util.asList(Action.USE))
+                                    ._postDuty_(Util.asList(new DutyBuilder()
+                                            ._action_(Util.asList(Action.NOTIFY))
+                                            ._constraint_(Util.asList(new ConstraintBuilder()
+                                                    ._leftOperand_(LeftOperand.ENDPOINT)
+                                                    ._operator_(BinaryOperator.DEFINES_AS)
+                                                    ._rightOperand_(
+                                                            new RdfResource(url, URI.create("xsd:anyURI")))
+                                                    .build()))
+                                            .build()))
+                                    .build()))
+                            .build();
+                    break;
+                }
+            case USAGE_LOGGING:
+                contractOffer = new ContractOfferBuilder()
+                        ._permission_(Util.asList(new PermissionBuilder()
+                                ._title_(Util.asList(new TypedLiteral("Usage Policy")))
+                                ._description_(Util.asList(new TypedLiteral("usage-logging")))
+                                ._action_(Util.asList(Action.USE))
+                                ._postDuty_(Util.asList(new DutyBuilder()
+                                        ._action_(Util.asList(Action.LOG))
+                                        .build()))
+                                .build()))
+                        .build();
+                break;
+
+            case USAGE_DURING_INTERVAL:
+                if (contractJson != null && !contractJson.equals("{}")) {
+                    final var jsonNode = getJsonNodeFromContract(contractJson);
+                    String fromDate = jsonNode.get("fromDate").asText();
+                    String toDate = jsonNode.get("toDate").asText();
+
+                    contractOffer = new ContractOfferBuilder()
+                            ._permission_(Util.asList(new PermissionBuilder()
+                                    ._title_(Util.asList(new TypedLiteral("Usage Policy")))
+                                    ._description_(Util.asList(new TypedLiteral("usage-during-interval")))
+                                    ._action_(Util.asList(Action.USE))
+                                    ._constraint_(Util.asList(new ConstraintBuilder()
+                                            ._leftOperand_(LeftOperand.POLICY_EVALUATION_TIME)
+                                            ._operator_(BinaryOperator.AFTER)
+                                            ._rightOperand_(new RdfResource(fromDate,
+                                                    URI.create("xsd:dateTimeStamp")))
+                                            .build(), new ConstraintBuilder()
+                                            ._leftOperand_(LeftOperand.POLICY_EVALUATION_TIME)
+                                            ._operator_(BinaryOperator.BEFORE)
+                                            ._rightOperand_(new RdfResource(toDate,
+                                                    URI.create("xsd:dateTimeStamp")))
+                                            .build()))
+                                    .build()))
+                            .build();
+                    break;
+                }
+            case USAGE_UNTIL_DELETION:
+                if (contractJson != null && !contractJson.equals("{}")) {
+                    final var jsonNode = getJsonNodeFromContract(contractJson);
+                    String startDate = jsonNode.get("startDate").asText();
+                    String endDate = jsonNode.get("endDate").asText();
+                    String deletionDate = jsonNode.get("deletionDate").asText();
+
+                    contractOffer = new ContractOfferBuilder()
+                            ._permission_(Util.asList(new PermissionBuilder()
+                                    ._title_(Util.asList(new TypedLiteral("Usage Policy")))
+                                    ._description_(Util.asList(new TypedLiteral("usage-until-deletion")))
+                                    ._action_(Util.asList(Action.USE))
+                                    ._constraint_(Util.asList(new ConstraintBuilder()
+                                            ._leftOperand_(LeftOperand.POLICY_EVALUATION_TIME)
+                                            ._operator_(BinaryOperator.AFTER)
+                                            ._rightOperand_(new RdfResource(startDate,
+                                                    URI.create("xsd:dateTimeStamp")))
+                                            .build(), new ConstraintBuilder()
+                                            ._leftOperand_(LeftOperand.POLICY_EVALUATION_TIME)
+                                            ._operator_(BinaryOperator.BEFORE)
+                                            ._rightOperand_(new RdfResource(endDate,
+                                                    URI.create("xsd:dateTimeStamp")))
+                                            .build()))
+                                    ._postDuty_(Util.asList(new DutyBuilder()
+                                            ._action_(Util.asList(Action.DELETE))
+                                            ._constraint_(Util.asList(new ConstraintBuilder()
+                                                    ._leftOperand_(LeftOperand.POLICY_EVALUATION_TIME)
+                                                    ._operator_(BinaryOperator.TEMPORAL_EQUALS)
+                                                    ._rightOperand_(new RdfResource(deletionDate,
+                                                            URI.create("xsd:dateTimeStamp")))
+                                                    .build()))
+                                            .build()))
+                                    .build()))
+                            .build();
+                    break;
+                }
+        }
+        return contractOffer;
+    }
+
+    /**
+     * @param contractJson the contract offer
+     * @return json node
+     */
+    private JsonNode getJsonNodeFromContract(String contractJson) throws JsonProcessingException {
+        final var mapper = new ObjectMapper();
+        return mapper.readTree(Objects.requireNonNull(contractJson));
     }
 }